Set 'role' field as required in Content and InputContent classes
The 'role' parameter has been set as a required field in the constructors of Content and InputContent classes. Previously, it was optional and could be null, but it now has "user" as a default value in Content and must be defined in InputContent. This change enforces that every instance of these classes will have a defined 'role', preventing possible null exceptions during runtime.