Update Content and InputContent classes, remove StartChatParams
Refactored the handling of Content and InputContent by removing InputContent class and absorbing functionalities in Content class. Introduced the concept of role in Content with factory constructors Content.user and Content.model. Also, removed the StartChatParams class. This refactoring simplifies the content and role management related to 'generative_ai_dart'.